==> notebooks/01-Scrape-RPM.ipynb <==
### Imports

*Note: This code runs in Python 2*

from bs4 import BeautifulSoup
import requests
from HTMLParser import HTMLParser
from datetime import datetime
import hashlib
import requests
import re

import numpy as np
import pandas as pd

class MLStripper(HTMLParser):
    def __init__(self):
        self.reset()
        self.strict = False
        self.convert_charrefs= True
        self.fed = []
    def handle_data(self, d):
        self.fed.append(d)
    def get_data(self):
        return ''.join(self.fed)

def strip_tags(html):
    s = MLStripper()
    s.feed(html)
    return s.get_data()


# In[58]:
from selenium import webdriver
import urllib2
#import MySQLdb

## Go to the ESPN site and get the data you want, which is the RPM info. 

season_RPM_df = []
for year in ['2014', '2015', '2016', '2017']:   
    #initialize data frame
    RPM_data = pd.DataFrame(columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))

    print str(datetime.now())
    for j in range(0, 12):

        url = "http://www.espn.com/nba/statistics/rpm/_/year/"+year+"/page/" + str(j) + "/sort/RPM"

        print url

        r1 = urllib2.urlopen(url)
        soup = BeautifulSoup(r1, 'html.parser')


        #get the even columns
        evens = soup.findAll('tr', {'class':'evenrow'})

        for i in range(0, len(evens)):
            name = evens[i].a.text
            name

            team = evens[i].find("td").find_next_sibling("td").find_next_sibling("td")
            team.text

            games_played = team.find_next_sibling("td")
            games_played.text

            mins_per_game = games_played.find_next_sibling("td")
            mins_per_game.text

            ORPM = mins_per_game.find_next_sibling("td")
            ORPM.text

            DRPM = ORPM.find_next_sibling("td")
            DRPM.text

            RPM = DRPM.find_next_sibling("td")
            RPM.text

            wins = RPM.find_next_sibling("td")
            wins.text

            record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

            temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
            frames = [RPM_data, temp_df]
            RPM_data = pd.concat(frames, ignore_index = True)


        #get the odd columns
        odds = soup.findAll('tr', {'class':'oddrow'})

        for i in range(0, len(odds)):
            name = odds[i].a.text
            name

            team = odds[i].find("td").find_next_sibling("td").find_next_sibling("td")
            team.text

            games_played = team.find_next_sibling("td")
            games_played.text

            mins_per_game = games_played.find_next_sibling("td")
            mins_per_game.text

            ORPM = mins_per_game.find_next_sibling("td")
            ORPM.text

            DRPM = ORPM.find_next_sibling("td")
            DRPM.text

            RPM = DRPM.find_next_sibling("td")
            RPM.text

            wins = RPM.find_next_sibling("td")
            wins.text

            record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

            temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
            frames = [RPM_data, temp_df]
            RPM_data = pd.concat(frames, ignore_index = True) #concatenate the odds and evens
        print len(RPM_data)
        # Include Season
        RPM_data['Season_Start_Year'] = '20'+ str(int(year[2:]) - 1)
        RPM_data['Season_End_Year'] = year

        # Remove duplicate first-last name
        RPM_dataset = RPM_data.drop_duplicates(subset=['names', 'Season_Start_Year', 'Season_End_Year'])
        
        
        season_RPM_df.append(RPM_dataset)
        

        
print str(datetime.now())

## Special Case: 2017-28 season (current season url)

#initialize data frame
RPM_data = pd.DataFrame(columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))

print str(datetime.now())
# 14 Pages of players unlike previous years
for j in [i+1 for i in range(0, 14)]:
    # First page doesn't list year
    if j == 1:
        url = "http://www.espn.com/nba/statistics/rpm/_/sort/RPM"
    # Current season has year of an underscore and not date year
    else:
        url = "http://www.espn.com/nba/statistics/rpm/_/page/"+ str(j) +"/sort/RPM"


    print url

    r1 = urllib2.urlopen(url)
    soup = BeautifulSoup(r1, 'html.parser')


    #get the even columns
    evens = soup.findAll('tr', {'class':'evenrow'})

    for i in range(0, len(evens)):
        name = evens[i].a.text
        name

        team = evens[i].find("td").find_next_sibling("td").find_next_sibling("td")
        team.text

        games_played = team.find_next_sibling("td")
        games_played.text

        mins_per_game = games_played.find_next_sibling("td")
        mins_per_game.text

        ORPM = mins_per_game.find_next_sibling("td")
        ORPM.text

        DRPM = ORPM.find_next_sibling("td")
        DRPM.text

        RPM = DRPM.find_next_sibling("td")
        RPM.text

        wins = RPM.find_next_sibling("td")
        wins.text

        record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

        temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
        frames = [RPM_data, temp_df]
        RPM_data = pd.concat(frames, ignore_index = True)


    #get the odd columns
    odds = soup.findAll('tr', {'class':'oddrow'})

    for i in range(0, len(odds)):
        name = odds[i].a.text
        name

        team = odds[i].find("td").find_next_sibling("td").find_next_sibling("td")
        team.text

        games_played = team.find_next_sibling("td")
        games_played.text

        mins_per_game = games_played.find_next_sibling("td")
        mins_per_game.text

        ORPM = mins_per_game.find_next_sibling("td")
        ORPM.text

        DRPM = ORPM.find_next_sibling("td")
        DRPM.text

        RPM = DRPM.find_next_sibling("td")
        RPM.text

        wins = RPM.find_next_sibling("td")
        wins.text

        record = [name, games_played.text, mins_per_game.text, ORPM.text, DRPM.text, RPM.text, wins.text]

        temp_df = pd.DataFrame([record],columns=('names', 'GP','MPG', 'ORPM', 'DRPM', 'RPM', 'WINS'))
        frames = [RPM_data, temp_df]
        RPM_data = pd.concat(frames, ignore_index = True) #concatenate the odds and evens
print len(RPM_data)
# Include Season
RPM_data['Season_Start_Year'] = '2017'
RPM_data['Season_End_Year'] = '2018'

# Remove duplicate first-last name
RPM_dataset = RPM_data.drop_duplicates(subset=['names', 'Season_Start_Year', 'Season_End_Year'])


season_RPM_df.append(RPM_dataset)

## Concat all together

RPM_df = pd.concat(season_RPM_df)

RPM_df.groupby(by=['Season_Start_Year', 'Season_End_Year']).count()

Duplicates often occur resulting in many more than the actual number of players (except the current season)

Drop duplicates of the subset player name, season to reduce to unique RPM player stats

RPM_df['names'].value_counts().head()

# Drop duplicates
RPM_df = RPM_df.drop_duplicates(subset=['names', 'Season_Start_Year', 'Season_End_Year'])

RPM_df.groupby(by=['Season_Start_Year', 'Season_End_Year']).count()

Sanity Check: Sort for highest RPM in the dataset

RPM_df.sort_values(by='RPM', ascending=False).head(7)

## Export

RPM_df.to_csv('player_RPM_stats.csv', index=False)
