==> notebooks/02-Join-Players.ipynb <==
import ruamel.yaml as yaml
import os
import sys
import pandas as pd
import numpy as np


NO_CONFIG_ERR_MSG = """No config file found. Root directory is determined by presence of "config.yaml" file."""

original_wd = os.getcwd()

# Number of times to move back in directory
num_retries = 10
for x in range(0, num_retries):
    # try to load config file
    try:
        with open("config.yaml", 'r') as stream:
            cfg = yaml.safe_load(stream)
    # If not found move back one directory level
    except FileNotFoundError:
        os.chdir('../')
        # If reached the max number of directory levels change to original wd and print error msg
        if x+1 == num_retries:
            os.chdir(original_wd)
            print(NO_CONFIG_ERR_MSG)

# Add directory to PATH
path = os.getcwd()

if path not in sys.path:
    sys.path.append(path)


## Load Data

RPM_df = pd.read_csv('data/interim/player_RPM_stats.csv')
box_score_df = pd.read_csv('data/raw/Box_Scores.csv')

## Joining RPM to players

Need to join on composite key made up of Player Name + Season

season_id appears to be calendar year the season started in 

Based off of the latest season being 2017 and not 2018

#### Steps:

1) RPM: Seperate names into first and last (start by splitting on a space)

2) Match season_id to either the calendar year the season ended or began. Map to RPM data

3) Join on First, Last, Season

box_score_df['season_id'].unique()

RPM_df.head()

# most number of whitespaces 
RPM_df['names'].str.split(' ').apply(len).max()

# Who has more than one space?

RPM_df['name_length_post_split'] = RPM_df['names'].str.split(' ').apply(len)
RPM_df.query("name_length_post_split > 2").names.value_counts().head()

# Mostly Jrs and some multiple names

What's the RPM of these players?

RPM_df.query("name_length_post_split > 2").RPM.describe()

first_name_to_search = 'Tim'
box_score_df[box_score_df.First_Name.str.contains(first_name_to_search)].head()

#### For now, will join to get RPM on a season/team/player level and fix this later if necessary

Split Names

RPM_df['First_Name'] = RPM_df['names'].str.split(' ').str.get(0)
RPM_df['Last_Name'] = RPM_df['names'].str.split(' ').str.get(-1)

Rename Starting Calendar year to season_id

RPM_df = RPM_df.rename(columns={'Season_Start_Year':'season_id'})

potential_keys = [
    'First_Name',
    'Last_Name',
    'season_id',
]

## Merge

RPM_player_data = (
    pd.merge(RPM_df, box_score_df, on=potential_keys, how='left')
    .drop_duplicates(subset=potential_keys) # Drop duplicate games, keep only first         
            )


# Check for nulls from box score data
RPM_player_data[RPM_player_data['Total_Rebounds'].isnull()].shape

RPM_player_data.head()

## Add Minutes Played Season Total to compute a team level average RPM weighted by minutes

RPM_player_data['Season_Total_Minutes'] = RPM_player_data['GP'] * RPM_player_data['MPG'] 
RPM_player_data['Season_Total_Minutes*RPM'] = RPM_player_data['Season_Total_Minutes'] \
* RPM_player_data['RPM']

RPM_player_data.head()

## Export all info for players

player_info_export_cols = ['DRPM',
 'GP',
 'MPG',
 'ORPM',
 'RPM',
 'Season_End_Year',
 'season_id',
 'WINS',
 'names',
 'name_length_post_split',
 'First_Name',
 'Last_Name',
 'Game_id',
 'Person_id',
 'Team_id',
 'Season_Total_Minutes',
 'Season_Total_Minutes*RPM']

RPM_player_data[player_info_export_cols].to_csv('data/interim/player-data-RPM-with-identifiers.csv', index=False)

keep_stats = [
 'DRPM',
 'ORPM',
 'RPM',
 'WINS',
]

team_minutes_weighted_avg_RPM = (RPM_player_data.groupby(by=['Team_id', 'season_id'])
 .sum()['Season_Total_Minutes*RPM'] / RPM_player_data.groupby(by=['Team_id', 'season_id'])
 .sum()['Season_Total_Minutes'])
team_minutes_weighted_avg_RPM = (team_minutes_weighted_avg_RPM.to_frame('Team_Minutes_Weighted_Avg_RPM')
                                 .reset_index())

team_RPM_total = (RPM_player_data.groupby(by=['Team_id', 'season_id'])
                  .sum()
                  .loc[:, keep_stats]
                  .add_prefix('Sum_')
                  .reset_index()
)
team_RPM_total.head()

team_level_RPM_stats = pd.merge(team_minutes_weighted_avg_RPM,
                               team_RPM_total,
                               on=['Team_id', 'season_id'])
team_level_RPM_stats['Team_id'] = team_level_RPM_stats['Team_id'].astype(int)

keep_cols = ['Team_id',
 'Season_id',
             'City',
 'Short_Name',
 'Nickname',]
team_mapping =(pd.read_csv('data/raw/Team_Mapping.csv')
               .loc[:, keep_cols]
               .rename(columns={'Season_id':'season_id'})
               # Remove leading '2' from season id
              .assign(season_id= lambda x: x['season_id'].astype(str).str.slice(1).astype(int)))
team_mapping.head()

## Adding team identifiers to data

team_level_RPM_stats = (pd.merge(team_mapping, 
                                team_level_RPM_stats,
                               on=['Team_id', 'season_id'])
                        .drop_duplicates()                                              
                       )
team_level_RPM_stats.head()

## Sorting teams high to low as a sanity check

team_level_RPM_stats.sort_values(by='Team_Minutes_Weighted_Avg_RPM', ascending=False).head(7)

team_level_RPM_stats.sort_values(by='Sum_WINS', ascending=False).head(7)

## Export data

team_level_RPM_stats.to_csv('data/interim/team_level_RPM_stats.csv', index=False)
