==> notebooks/05-Updated-model.ipynb <==
import ruamel.yaml as yaml
import os
import sys
import pandas as pd
import numpy as np
import xgboost as xgb
from sklearn import preprocessing
from sklearn.cross_validation import train_test_split

from sklearn.cross_validation import *
from sklearn.grid_search import GridSearchCV

NO_CONFIG_ERR_MSG = """No config file found. Root directory is determined by presence of "config.yaml" file."""

original_wd = os.getcwd()

# Number of times to move back in directory
num_retries = 10
for x in range(0, num_retries):
    # try to load config file
    try:
        with open("config.yaml", 'r') as stream:
            cfg = yaml.safe_load(stream)
    # If not found move back one directory level
    except FileNotFoundError:
        os.chdir('../')
        # If reached the max number of directory levels change to original wd and print error msg
        if x+1 == num_retries:
            os.chdir(original_wd)
            print(NO_CONFIG_ERR_MSG)

# Add directory to PATH
path = os.getcwd()

if path not in sys.path:
    sys.path.append(path)


## Load and Process Data

df = pd.read_csv('data/processed/all_game_data.csv')

One Hot Encode the season

pd.options.display.max_columns=150
df.head()

id_cols = ['Game_id',
           'season_id',
           'Team_id',
           'Season',
           'Date',
           'Team_Short'
          ]
id_cols += df.filter(regex='player\d$').columns.values.tolist()

# Remove any player after 8th most played player
remove_features = df.iloc[:2, df.columns.get_loc('player9'):].columns.values.tolist()
# Keep Total RPM and Total Wins
remove_features = [feat for feat in remove_features if feat not in ['Tot_WINS', 'Tot_RPM']]
# Remove One Hot Encode Cols
remove_features += ['Team_Name',
           'Home_or_Away', 
                   'day_of_week']

## One hot encode and set X and y

## Cleaning Nulls

possible_feats = df[[col for col in df.columns.values.tolist() if col not in remove_features]]
possible_feats.shape

## Since the Wins entering game likely has quite a bit to do with TV rating and ticket price, NaNs will be dropped.
## Possibly will do two models, one with and one without to see which is the beter performing since this drops half the data
## Player stats will be mean imputed since there's not a huge number of nulls

from src.utils.pandas_utils import null_column_report_df
from src.utils.custom_transformers import DFImputer
null_column_report_df(possible_feats)

## Dropping Nulls

drop_cols = ['Season', 'day_of_week', 'Ticket_Value', 'Wins_Entering_Gm', 'Losses_Entering_Gm', 'WIN_Perc']
drop_cols_keep_win_data = ['Season', 'day_of_week', 'Ticket_Value']
drop_cols_win_data_cols = ['Wins_Entering_Gm', 'Losses_Entering_Gm', 'WIN_Perc']
df_dropped_null_wins_data = df.dropna(subset=drop_cols)
df_kept_null_wins_data = df.dropna(subset=drop_cols_keep_win_data)

## Mean Imputing

mean_impute_cols = [col for col 
                    in null_column_report_df(possible_feats).Column.values.tolist() 
                    if col not in drop_cols]
print(mean_impute_cols[:3])
df_kept_null_wins_data[mean_impute_cols] = DFImputer().fit_transform(df_kept_null_wins_data[mean_impute_cols])
df_dropped_null_wins_data[mean_impute_cols] = DFImputer().fit_transform(df_dropped_null_wins_data[mean_impute_cols])

## One Hot Categoric Data

target_cols = [
    'Ticket_Value', 'TV_Rate'
]

def one_hot_df(games_df):
    from functools import reduce
    season_one_hot = pd.get_dummies(df['Season']).add_prefix('Season_')
    day_of_week_one_hot = pd.get_dummies(df['day_of_week']).add_prefix('Day_Of_Week')
    team_one_hot = pd.get_dummies(df['Team_Name']).add_prefix('Team_')
    frames = [season_one_hot, day_of_week_one_hot, team_one_hot] # add arguments as necessary to the read_csv method
    one_hot_features_df = reduce(lambda left,right: pd.merge(left,right,
                                                             left_index=True,
                                                            right_index=True), 
                                 frames)

    return one_hot_features_df

## Concat_together

features = [col for col in df.columns.values.tolist() if col not in target_cols+id_cols+remove_features]

X = pd.merge(one_hot_df(df_dropped_null_wins_data),
             df_dropped_null_wins_data[features],
             left_index=True,
             right_index=True)
Y_Tickets = df_dropped_null_wins_data['Ticket_Value']
Y_TV = df_dropped_null_wins_data['TV_Rate']

X_kept_win_data = pd.merge(one_hot_df(df_kept_null_wins_data), 
                          df_kept_null_wins_data[features],
                         left_index=True,
                         right_index=True)
X_kept_win_data = X_kept_win_data.drop(columns=drop_cols_win_data_cols)
Y_kept_win_data_Tickets = df_kept_null_wins_data['Ticket_Value']
Y_kept_win_data_TV = df_kept_null_wins_data['TV_Rate']

## Sanity Check

null_column_report_df(X)

X.shape

null_column_report_df(X_kept_win_data)

X_kept_win_data.shape

## Export Data

pd.concat([X, Y_Tickets, Y_TV], axis=1).to_csv('data/processed/modelling_data_dropped_null_wins.csv', index=False)
pd.concat([X_kept_win_data, Y_kept_win_data_Tickets, Y_kept_win_data_TV], axis=1).to_csv('data/processed/modelling_data.csv', index=False)

## CV Score with XGB regressor

xgb_model = xgb.XGBRegressor()

from sklearn.model_selection import cross_val_score

%%time
cv_score_tickets = cross_val_score(xgb_model, X=X, y=Y_Tickets, cv=10,
                                                scoring='neg_mean_squared_error')                                   
cv_score_tv = cross_val_score(xgb_model, X=X, y=Y_TV, cv=10,
                                                scoring='neg_mean_squared_error')
                              

cv_score_tickets.mean()

cv_score_tv.mean()

%%time
cv_score_kept_win_data_tickets = cross_val_score(xgb_model, X=X_kept_win_data, y=Y_kept_win_data_Tickets, cv=10,
                                                scoring='neg_mean_squared_error')
cv_score_kept_win_data_tv = cross_val_score(xgb_model, X=X_kept_win_data, y=Y_kept_win_data_TV, cv=10,
                                                                                            scoring='neg_mean_squared_error')


cv_score_kept_win_data_tickets.mean()

cv_score_kept_win_data_tv.mean()

# The Data with the wins dropped slightly outperformed the data with the wins kept. Because it's half the size, it will be used

from sklearn.model_selection import RandomizedSearchCV

#brute force scan for all parameters, here are the tricks
#usually max_depth is 6,7,8
#learning rate is around 0.05, but small changes may make big diff
#tuning min_child_weight subsample colsample_bytree can have 
#much fun of fighting against overfit 
#n_estimators is how many round of boosting
#finally, ensemble xgboost with multiple seeds may reduce variance
parameters = {'nthread':[4], #when use hyperthread, xgboost may become slower
              'learning_rate': [.001, 0.05, .01], #so called `eta` value
              'max_depth': [2, 5, 10, 20],
              'min_child_weight': [11],
              'silent': [1],
              'subsample': [.2, .5, 0.8],
              'colsample_bytree': [.2, .5, 0.8],
              'n_estimators': [5, 50, 500], #number of trees, change it to 1000 for better results
              'missing':[-999],
              'seed': [42]}


clf = RandomizedSearchCV(xgb_model, parameters, n_jobs=6, 
                   cv=5, verbose=2, refit=True)

%%time
clf.fit(X, Y_Tickets)

#trust your CV!
best_parameters, score, _ = max(clf.grid_scores_, key=lambda x: x[1])
print('Score:', score)
for param_name in sorted(best_parameters.keys()):
    print("%s: %r" % (param_name, best_parameters[param_name]))

# test_probs = clf.predict_proba(test[features])[:,1]

# sample = pd.read_csv('../input/sample_submission.csv')
# sample.QuoteConversion_Flag = test_probs
# sample.to_csv("xgboost_best_parameter_submission.csv", index=False)

best_parameters

## Repeating but with Ticket Info

#brute force scan for all parameters, here are the tricks
#usually max_depth is 6,7,8
#learning rate is around 0.05, but small changes may make big diff
#tuning min_child_weight subsample colsample_bytree can have 
#much fun of fighting against overfit 
#n_estimators is how many round of boosting
#finally, ensemble xgboost with multiple seeds may reduce variance
tv_parameters = {'nthread':[4], #when use hyperthread, xgboost may become slower
              'learning_rate': [.001, 0.05, .01], #so called `eta` value
              'max_depth': [2, 5, 10, 20],
              'min_child_weight': [11],
              'silent': [1],
              'subsample': [.2, .5, 0.8],
              'colsample_bytree': [.2, .5, 0.8],
              'n_estimators': [5, 50, 500], #number of trees, change it to 1000 for better results
              'missing':[-999],
              'seed': [42]}


tv_clf = RandomizedSearchCV(xgb_model, tv_parameters, n_jobs=6, 
                   cv=5, verbose=2, refit=True)

%%time
tv_clf.fit(X, Y_TV)

#trust your CV!
tv_best_parameters, tv_score, _ = max(tv_clf.grid_scores_, key=lambda x: x[1])
print('Score:', tv_score)
for param_name in sorted(tv_best_parameters.keys()):
    print("%s: %r" % (param_name, tv_best_parameters[param_name]))

# test_probs = clf.predict_proba(test[features])[:,1]

# sample = pd.read_csv('../input/sample_submission.csv')
# sample.QuoteConversion_Flag = test_probs
# sample.to_csv("xgboost_best_parameter_submission.csv", index=False)

tv_best_parameters
